Rename add_heal.ts to heal.ts
